+++ src/ShellCheck/Parser.hs
@@ -3580,8 +3580,8 @@ readScriptFile sourced = do
     verifyShebang pos s = do
         case isValidShell s of
             Just True -> return ()
-            Just False -> parseProblemAt pos ErrorC 1071 "ShellCheck only supports sh/bash/dash/ksh/'busybox sh'/irix-bsh/irix-jsh/irix-sh/irix-ksh scripts. Sorry!"
-            Nothing -> parseProblemAt pos ErrorC 1008 "This shebang was unrecognized. ShellCheck only supports sh/bash/dash/ksh/'busybox sh'/irix-bsh/irix-jsh/irix-sh/irix-ksh. Add a 'shell' directive to specify."
+            Just False -> parseProblemAt pos ErrorC 1071 "ShellCheck only supports sh/bash/dash/ksh/'busybox sh'/irix-bsh/irix-jsh/irix-sh/irix-ksh/irix-dtksh scripts. Sorry!"
+            Nothing -> parseProblemAt pos ErrorC 1008 "This shebang was unrecognized. ShellCheck only supports sh/bash/dash/ksh/'busybox sh'/irix-bsh/irix-jsh/irix-sh/irix-ksh/irix-dtksh. Add a 'shell' directive to specify."
 
     isValidShell s =
         let good = null s || any (`isPrefixOf` s) goodShells
@@ -3607,7 +3607,9 @@ readScriptFile sourced = do
         "irix-bsh",
         "irix-jsh",
         "irix-sh",
-        "irix-ksh"
+        "irix-ksh",
+        "dtksh",
+        "irix-dtksh"
         ]
     badShells = [
         "awk",